+++ src/hw/hw_def.h
@@ -6,7 +6,7 @@
 #include QMK_KEYMAP_CONFIG_H
 
 
-#define _DEF_FIRMWATRE_VERSION      "V251010R5"  // V251010R5: LED 모듈 분할 및 WS2812 더티 플래그
+#define _DEF_FIRMWATRE_VERSION      "V251010R4"  // V251010R4: 호스트 LED 지연 적용 및 RGBlight 재합성 1회화
 #define _DEF_BOARD_NAME             "BARAM-QMK-H7S-FW"
